Add ATA PIO support
- Introduced memory management functions: krealloc and kcalloc in memory.c.
- Modified existing memory management functions to improve memory allocation and deallocation.
- Updated the main kernel entry point to clear the BSS segment.
- Refactored IDE controller code to use ATA standards, including new definitions and functions for ATA operations.
- Improved error handling and added timeout mechanisms for ATA device interactions.
- Cleaned up Makefiles to replace 'mrproper' with 'clean'.
- Removed unused IDE Bus Master base retrieval function from PCI handling.
- Added delay function in PIT for better timing control.
- Updated RTC timestamp function to return a 64-bit value for improved time representation.

+++ source/include/ClassiX/io.h
@@ -17,9 +17,9 @@
 #define nop()								asm volatile ("nop")
 #define pause()								asm volatile ("pause")
 
-#define out8(port, data)					asm volatile ("outb %%al, %%dx"::"a"(data), "d"(port))
-#define out16(port, data)					asm volatile ("outw %%ax, %%dx"::"a"(data), "d"(port))
-#define out32(port, data)					asm volatile ("outl %%eax, %%dx"::"a"(data), "d"(port))
+#define out8(port, data)					asm volatile ("outb %%al, %%dx"::"a"((uint8_t) (data)), "d"(port))
+#define out16(port, data)					asm volatile ("outw %%ax, %%dx"::"a"((uint16_t) (data)), "d"(port))
+#define out32(port, data)					asm volatile ("outl %%eax, %%dx"::"a"((uint32_t) (data)), "d"(port))
 
 #define in8(port) ({										\
 	uint8_t _data;											\
